Tests: Improve test cases for `mqttwarn.commands`
