+++ tests/util.py
@@ -1,9 +1,12 @@
 # -*- coding: utf-8 -*-
 # (c) 2018-2021 The mqttwarn developers
+import shlex
 import threading
+from unittest.mock import patch
 
 import mqttwarn
 import paho
+from mqttwarn.commands import run as run_command
 from mqttwarn.configuration import load_configuration
 from mqttwarn.core import bootstrap, load_services, on_message, start_workers
 from paho.mqtt.client import MQTTMessage
@@ -61,3 +64,12 @@ def mqtt_process(mqttc: paho.mqtt.client.Client, loops=2):
     for _ in range(loops):
         mqttc.loop()
     delay()
+
+
+def invoke_command(capfd, command):
+    if not isinstance(command, list):
+        command = shlex.split(command)
+    with patch("sys.argv", command):
+        run_command()
+    output = capfd.readouterr()
+    return output.out, output.err
